Bit of casting to remove a few more warnings

+++ libsrc/c128/ellipsevdc.c
@@ -29,10 +29,10 @@ void ellipsevdc(int XC, int YC, int A, int B)
 
     while (XBB2 <= YAA2) /* draw octant from top to top right */
     {
-      setpixvdc(XC+X,YC+Y);
-      setpixvdc(XC+X,YC-Y);
-      setpixvdc(XC-X,YC+Y);
-      setpixvdc(XC-X,YC-Y);
+      setpixvdc((int)(XC+X),(int)(YC+Y));
+      setpixvdc((int)(XC+X),(int)(YC-Y));
+      setpixvdc((int)(XC-X),(int)(YC+Y));
+      setpixvdc((int)(XC-X),(int)(YC-Y));
       X += 1;
       XBB2 += BB2;
       ErrVal += XBB2-BB;
@@ -53,10 +53,10 @@ void ellipsevdc(int XC, int YC, int A, int B)
 
     while (XBB2 > YAA2)  /* draw octant from right to top right */
     {
-      setpixvdc(XC+X,YC+Y);
-      setpixvdc(XC+X,YC-Y);
-      setpixvdc(XC-X,YC+Y);
-      setpixvdc(XC-X,YC-Y);
+      setpixvdc((int)(XC+X),(int)(YC+Y));
+      setpixvdc((int)(XC+X),(int)(YC-Y));
+      setpixvdc((int)(XC-X),(int)(YC+Y));
+      setpixvdc((int)(XC-X),(int)(YC-Y));
       Y += 1;
       YAA2 += AA2;
       ErrVal += YAA2-AA;
